20040929
git-svn-id: svn://sourceware.org/svn/prelink@117 94c539fb-cf18-0410-b60f-edeeb537fa16

+++ src/dso.c
@@ -1620,11 +1620,8 @@ close_dso (DSO *dso)
 {
   int rdwr = dso_is_rdwr (dso);
 
-  if (rdwr)
-    {
-      assert (dso->temp_filename != NULL);
-      unlink (dso->temp_filename);
-    }
+  if (rdwr && dso->temp_filename != NULL)
+    unlink (dso->temp_filename);
   close_dso_1 (dso);
   return 0;
 }
